requirements.txt
streamlit
pandas
plotly
firebase-admin
pyrebase4
python-dotenv
streamlit-cookies-manager

